move va_awards_composite to own file, added YARD docs
