+++ lib/debt_management_center/models/va_awards_composite.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module DebtManagementCenter
+  class VaAwardsComposite
+    include Virtus.model
+    attribute :name, String
+    attribute :amount, String
+    attribute :veteran_or_spouse, String
+  end
+end
